training/m1_ray/requirements_ray.txt
pandas
scikit-learn
xgboost
mlflow
pyyaml
scipy
ray[train]==2.35.0
boto3
pyarrow
